+++ edison-mongo/src/test/java/de/otto/edison/mongo/configuration/MongoPropertiesTest.java
@@ -8,27 +8,10 @@ import static org.junit.Assert.assertThat;
 public class MongoPropertiesTest {
 
     @Test
-    @SuppressWarnings("deprecation")
-    public void shouldReturnPasswordWhenDeprecatedPasswordFieldIsUsed() throws Exception {
-        MongoProperties props = new MongoProperties();
-        props.setPasswd("somePassword");
-        assertThat(props.getPassword(), is("somePassword"));
-    }
-
-    @Test
-    public void shouldReturnPasswordWhenNewPasswordFieldIsUsed() throws Exception {
+    public void shouldReturnPassword() throws Exception {
         MongoProperties props = new MongoProperties();
         props.setPassword("somePassword");
         assertThat(props.getPassword(), is("somePassword"));
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    public void shouldPreferNewPasswordOverDeprecatedOne() throws Exception {
-        MongoProperties props = new MongoProperties();
-        props.setPasswd("someDeprecatedPassword");
-        props.setPassword("someNewPassword");
-        assertThat(props.getPassword(), is("someNewPassword"));
     }
 
 }